requirements.txt
python-dotenv==1.0.1
ccxt==4.3.89
pandas==2.2.2
pandas-ta==0.3.14b0
numpy==1.26.4
feedparser==6.0.11
requests==2.32.3
openai==1.35.14
praw==7.7.1
tweepy==4.14.0
